Add files via upload
Program accepts 2 user inputs that act as lower and upper bounds. 
Program considers bounds as inclusive. 
Bounds are then run through a loop, where they are multiplied by [1-20). 
If a result (outcome of a bound * a number [1-20) is NOT divisible by 2 with no remainder: console prints 'buzz'
if a result (outcome of a bound * a number [1-20) IS divisible by 2 with no remainder: console prints 'fizz'